chore: minor cleaning

+++ Bouncing Mushroom/src/Game.hpp
@@ -19,7 +19,7 @@ namespace Daklit
 	private:
 		sf::Texture m_mushroomTexture;
 		sf::Sprite m_mushroom;
-		sf::Vector2i m_increment;
+		sf::Vector2f m_increment;
 
 		sf::Clock m_clock;
 		sf::Time m_elapsed;
